Move unique validator to a separate NaturalValidators class
